update nlp tests
