+++ nlp/tests/test_database.py
@@ -34,73 +34,6 @@ log = getLogger(__name__)
 ENTITIES = {"PERSON", "PLACE", "TIME"}
 
 
-class Config:
-    """minimal class for setting up an in memory db for this test"""
-
-    def __init__(self):
-        self.DB_URI = "postgresql+psycopg2://postgres:hardpass123@localhost:5432/nlp"
-        self.DEBUG = False  # outputs all activity
-
-
-def fill_db(engine: Engine, training_dir: str):
-    """Loads database with files found in base_training_data"""
-    log.info("Filling the DB with initial training data")
-    training_data = list()
-    for file in os.scandir(training_dir):
-        if os.path.isfile(file) and file.name.endswith(".json"):
-            with open(file.path, "r") as f:
-                json_file = json.load(f)
-                for entry in json_file:
-                    training_data.append(entry)
-    to_commit = list()
-    for citation in training_data:
-        citation_id = str(uuid4())
-        content = citation.get("content")
-        entities = citation.get("entities")
-        annotated_citation = AnnotatedCitation(text=content, id=citation_id)
-        to_commit.append(annotated_citation)
-        entity_list = [
-            Entity(
-                id=str(uuid4()),
-                start_char=e[0],
-                stop_char=e[1],
-                type=e[2],
-                annotated_citation=annotated_citation,
-            )
-            for e in entities
-        ]
-        to_commit.extend(entity_list)
-
-    log.info(f"Initializing DB with {len(to_commit)} objects")
-    with Session(engine, future=True) as session:
-        session.add_all(to_commit)
-        session.commit()
-
-
-def reset_db(db):
-    Base.metadata.drop_all(db._engine)
-    Base.metadata.create_all(db._engine)
-    fill_db(db._engine, db._config.TRAIN_DIR)
-
-
-@pytest.fixture
-def config():
-    return Config()
-
-
-@pytest.fixture
-def db(config) -> Database:
-    root = os.getcwd()
-    if root.endswith("nlp"):
-        config.TRAIN_DIR = root + "/tests/test_train_dir"
-    elif root.endswith("tests"):
-        config.TRAIN_DIR = root + "/test_train_dir"
-    config.debug = True
-    db = Database(config)
-    reset_db(db)
-    return db
-
-
 event_base_dict = {
     "transaction_id": "2aab2a27-6b61-47d3-b5ba-a5057d0a880c",
     "app_version": "0.0.1-test",
@@ -115,9 +48,8 @@ def citation_added():
     payload = CitationAddedPayload(
         id="efb1266e-4b79-437e-9fdb-b0e62eb2de3b",
         text=text,
-        tags=[],
         summary_id="ec0c7955-c66e-44ae-82aa-9338d78291c5",
-        meta="some info here",
+        meta_id="d4640c23-e491-4b17-a01b-2bb5e0ea0662",
     )
     return CitationAdded(**event_base_dict, payload=payload)
 
@@ -185,7 +117,7 @@ def place_added():
         citation_end=35,
         longitude=12.4964,
         latitude=41.9028,
-        geoshape=None,
+        geo_shape=None,
     )
     return PlaceAdded(**event_base_dict, payload=payload)
